Fixed massive, yet small bug

+++ c++/3dgeom.h
@@ -2,5 +2,5 @@
 #include <glm\glm.hpp>
 #include <vector>
 
-void createNonIndexedCube(GLfloat*, int sizex, int sizey, int sizez);
+void createNonIndexedCube(GLfloat*, float sizex, float sizey, float sizez);
 void addTriangleToVector(std::vector<glm::vec3>* vec, glm::vec3 p1, glm::vec3 p2, glm::vec3 p3);
